Refactor slides spacing for proper placeholder styling
