+++ client/my-sites/themes/collections/showcase-theme-collection.tsx
@@ -4,8 +4,8 @@ import { useQueryThemes } from 'calypso/components/data/query-themes';
 import Theme from 'calypso/components/theme';
 import ThemeCollection from 'calypso/components/theme-collection';
 import ThemeCollectionItem from 'calypso/components/theme-collection/theme-collection-item';
+import ThemeCollectionPlaceholder from 'calypso/components/theme-collection/theme-collection-placeholder';
 import { ThemeBlock } from 'calypso/components/themes-list';
-import ThemeCollectionPlaceholder from 'calypso/my-sites/themes/collections/theme-collection-placeholder';
 import { ThemeCollectionsLayoutProps } from 'calypso/my-sites/themes/collections/theme-collections-layout';
 import {
 	ThemesQuery,
